% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusGapGen.R
\name{clusGapGen.plot}
\alias{clusGapGen.plot}
\title{clusGapGen.plot}
\usage{
clusGapGen.plot(clusGapRes, divBySd = FALSE, main = "")
}
\arguments{
\item{clusGapRes, }{a list returned by a call of function clusGapGen}

\item{divBySd, }{logical, if TRUE plot for the standardize version of the gap}

\item{main, }{the main title to the plots}
}
\description{
clusGapGen.plot
}
