% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusGapGen.R
\name{clusGapGen}
\alias{clusGapGen}
\title{clusGapGen}
\usage{
clusGapGen(
  x,
  algo,
  index,
  modelH0,
  K.max,
  B = 100,
  transformData = identity,
  modelH0Param = function(y) list(),
  low = TRUE,
  verbose = interactive(),
  ...
)
}
\arguments{
\item{x}{the data, can be a data frame or a matrix}

\item{algo, }{a clustering algorithm function taking the prepared data and
a number of clusters as arguments}

\item{index, }{a function taking a clustering vector and the prepared data
which returns the value of a validity index. Index can also be a list of such
functions to obtain results for different indices.
For coherence with the originally proposed Gap Statistic in Tibshirani et al's
a LOWER value of the validity index implies a better clustering quality, so
indices such as average silhouette width should be added a minus sign.
This can be changed by setting the argument low=FALSE.}

\item{modelH0, }{a function which takes as argument at least the data x,
parameters estimated from the data, and further arguments in ...}

\item{K.max, }{number of different clusters for which the index should be
evaluated.}

\item{B, }{the number of bootstraps sample.}

\item{transformData, }{a function which takes the data x as argument and
processed it for clustering}

\item{modelH0Param, }{a function which takes as argument the data x and returns
a list of modelH0 parameters with matching names.}

\item{low, }{logical, if FALSE a HIGHER value of the index or the indices in the
user provided list implies a better clustering quality}
}
\value{
if index is just one function a list with components:
Indks, the bootstrap validity plots
Ind, the validity plot corresponding to the data
E.Ind, the sample mean of the bootstrap validity plots
gap, the calibrated validity plot, difference between E.Ind and Ind
gapHen, the gap divided by the standard deviation of the bootstraps validity plots
SE.sim, the sample standard error of the bootstrap validity plots with a correction
term for bootstrap estimation
SE, the sample standard error
if index is a list of index then each of the components above are lists with values
for each index
}
\description{
clusGapGen
}
