% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{lower.to.upper.tri.inds}
\alias{lower.to.upper.tri.inds}
\title{Return indices to *permute*  "dissimilarity" / "dist" entries for C (ex-Fortran) code setup}
\usage{
lower.to.upper.tri.inds(n)
}
\arguments{
\item{n}{"Size" = number of objects, underlying the dist/dissimilarity
used in  ./agnes.q, ./clara.q,  ./diana.q  und ./pam.q :
*somewhat* related to Matrix:::indTri()}
}
\description{
Currently always used as:
  n <- attr(x, "Size")
  dv <- x[lower.to.upper.tri.inds(n)]
-->> FIXME: eventually do the above directly in C
}
