% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusGap.R
\name{clusGap}
\alias{clusGap}
\title{clusGap}
\usage{
clusGap(
  x,
  FUNcluster,
  K.max,
  B = 100,
  d.power = 1,
  spaceH0 = c("scaledPCA", "original"),
  verbose = interactive(),
  parallel = FALSE,
  future_plan = "multisession",
  ...
)
}
\arguments{
\item{...}{}

\item{plan}{plan for resolving parallel futures (e.g. how should certain parallel calculations be performed).
See \code{future::plan} for acceptable options.  Default: "multisession"}
}
\description{
clusGap
}
